+++ paths/getahaircut.tsc
@@ -1,0 +1,10 @@
+-0.8	8.8
+-8.488888888888889	-2.0000000000000018
+-28.044444444444444	-18.133333333333333
+-23.866666666666667	-23.6
+-19.68888888888889	-29.06666666666667
+16.31111111111111	-31.88888888888889
+24.266666666666666	-24.0
+32.22222222222222	-16.11111111111111
+24.0	7.822222222222223
+23.866666666666667	23.733333333333334
