+++ paths/hendryneedsahaircut.tsv
@@ -1,0 +1,7 @@
+0.26666666666666666	0.26666666666666666
+8.400000000000002	-7.688888888888889
+15.997840878790877	-39.97444871043206
+24.666666666666668	-23.6
+18.666666666666668	22.933333333333334
+30.977777777777778	9.73333333333333
+34.13333333333333	26.4
